test user model use rspec,factory_girl,faker

+++ app/controllers/upload_photos_controller.rb
@@ -103,9 +103,7 @@ class UploadPhotosController < ApplicationController
   end
   
   def delPhoto(filename)
-    if !filename.empty?
-      FileUtils.rm($uploadFolder+"/"+filename)
-    end
+      FileUtils.rm($uploadFolder+"/"+filename) unless filename.empty?
   end
   
   def checkFileType(fileName)
